+++ R/build_urls.R
@@ -1,12 +1,11 @@
 #' @noRd
 build_wb_url <- function(base_url, indicator, path_list, query_list) {
-
   url_path <- unlist(path_list)
   url_path <- url_path[!is.na(url_path)]
 
   query_list <- query_list[!is.na(query_list)]
 
-  if(missing(indicator)) {
+  if (missing(indicator)) {
     out_url <- httr::modify_url(base_url, path = url_path, query = query_list)
     return(out_url)
   }
@@ -14,21 +13,22 @@ build_wb_url <- function(base_url, indicator, path_list, query_list) {
   indicator_path <- wb_api_parameters$indicator
   indicator_path <- paste0(indicator_path, "/", indicator)
 
-  if(is.null(names(indicator))) names(indicator_path) <- indicator
-  else names(indicator_path) <- names(indicator)
+  if (is.null(names(indicator))) {
+    names(indicator_path) <- indicator
+  } else {
+    names(indicator_path) <- names(indicator)
+  }
 
-  out_url <- sapply(indicator_path, FUN = function(ind) {
+  out_url <- vapply(indicator_path, FUN = function(ind) {
     url_path <- c(url_path, ind)
     httr::modify_url(base_url, path = url_path, query = query_list)
-  }
-  )
+  }, FUN.VALUE = character(1))
 
   out_url
 }
 
 #' @noRd
 build_get_url <- function(end_point, lang) {
-
   base_url <- wb_api_parameters$base_url
 
   path_list <- list(
@@ -38,9 +38,9 @@ build_get_url <- function(end_point, lang) {
   )
 
   query_list <- list(
-   # per_page = wb_api_parameters$per_page,
+    # per_page = wb_api_parameters$per_page,
     per_page = ifelse(end_point == "indicator", 500, wb_api_parameters$per_page),
-    format   = wb_api_parameters$format
+    format = wb_api_parameters$format
   )
 
 
@@ -55,69 +55,90 @@ build_get_url <- function(end_point, lang) {
 
 #' @noRd
 fetch_wb_url <- function(url_string, indicator) {
-
   return_json <- fetch_wb_url_content(url_string = url_string, indicator = indicator)
 
   return_list <- jsonlite::fromJSON(return_json, simplifyVector = FALSE)
 
   if ("message" %in% names(return_list[[1]])) {
-
     message_list <- return_list[[1]]$message[[1]]
 
-    stop(sprintf("World Bank API request failed for indicator %s The following message was returned from the server\nid: %s\nkey: %s\nvalue: %s\n\nfailed request:\n%s",
-                 indicator,
-                 message_list$id,
-                 message_list$key,
-                 message_list$value,
-                 url_string),
-         call. = FALSE)
-
+    stop(
+      sprintf(
+        "World Bank API request failed for indicator %s The following message was returned from the server\nid: %s\nkey: %s\nvalue: %s\n\nfailed request:\n%s",
+        indicator,
+        message_list$id,
+        message_list$key,
+        message_list$value,
+        url_string
+      ),
+      call. = FALSE
+    )
   }
 
   n_pages <- return_list[[1]]$pages
 
-  if (n_pages == 0) return(NA) # a blank data frame will be returned to the user
+  if (n_pages == 0) {
+    return(NA)
+  } # a blank data frame will be returned to the user
 
-  return_list <- jsonlite::fromJSON(return_json,  flatten = TRUE)
+  return_list <- jsonlite::fromJSON(return_json, flatten = TRUE)
 
   lastUpdated <- return_list[[1]]$lastupdated
 
   if (n_pages > 1) {
-
     page_list <- lapply(1:n_pages, FUN = function(page) {
-
       if (page == 1) {
-
         return_list[[2]]
-
       } else {
-
         page_url <- paste0(url_string, "&page=", page)
 
         page_return_json <- fetch_wb_url_content(url_string = page_url)
-        page_return_list <- jsonlite::fromJSON(page_return_json,  flatten = TRUE)
+        page_return_list <- jsonlite::fromJSON(page_return_json, flatten = TRUE)
         page_df <- page_return_list[[2]]
-
       }
-    }
-    ) # end lapply
+    }) # end lapply
 
     return_df <- do.call("rbind", page_list)
-
   } else { # only one page
 
     return_df <- return_list[[2]]
-
   }
 
   return_df$lastUpdated <- lastUpdated
   return_df
+}
 
+#' @noRd
+# retries transient failures (connection timeouts/resets and 5xx server
+# errors) up to `max_attempts` times with exponential backoff. 4xx client
+# errors are not retried since they will not succeed on a subsequent
+# attempt (e.g. the "Bad Request" footnote issue handled separately below)
+wb_api_get <- function(url_string, ua, max_attempts = 5) {
+  get_return <- tryCatch(
+    httr::RETRY(
+      "GET", url_string, ua, httr::timeout(20),
+      times = max_attempts,
+      terminate_on = 400:499,
+      quiet = TRUE
+    ),
+    error = function(e) e
+  )
+
+  if (inherits(get_return, "error")) {
+    stop(
+      sprintf(
+        "World Bank API request failed after %s attempts\nmessage: %s\nurl: %s",
+        max_attempts, conditionMessage(get_return), url_string
+      ),
+      call. = FALSE
+    )
+  }
+
+  get_return
 }
 
 #' @noRd
 fetch_wb_url_content <- function(url_string, indicator) {
-
   indicator <- if_missing(indicator)
 
   # move this to data-raw eventually
@@ -125,7 +146,7 @@ fetch_wb_url_content <- function(url_string, indicator) {
 
   # add api_token here if/when that is supported
 
-  get_return <- httr::GET(url_string, ua, httr::timeout(20))
+  get_return <- wb_api_get(url_string, ua)
 
   # there is a known issue with some sources without metadata failing when
   # the footnote field is requested. Since we can't know ahead of time
@@ -137,7 +158,7 @@ fetch_wb_url_content <- function(url_string, indicator) {
 
     if (error_status$reason == "Bad Request" & grepl(footnote_pattern, url_string)) {
       url_string_retry <- gsub(footnote_pattern, "footnote=n", url_string)
-      get_return_retry <- httr::GET(url_string_retry, ua, httr::timeout(20))
+      get_return_retry <- wb_api_get(url_string_retry, ua)
 
       # if this one returns successfully, then replace the original
       if (!httr::http_error(get_return_retry)) get_return <- get_return_retry
@@ -146,13 +167,17 @@ fetch_wb_url_content <- function(url_string, indicator) {
 
   # throw error if still returns error
   if (httr::http_error(get_return)) {
-    stop(sprintf("World Bank API request failed for indicator %s\nmessage: %s\ncategory: %s\nreason: %s \nurl: %s",
-                 indicator,
-                 error_status$message,
-                 error_status$category,
-                 error_status$reason,
-                 url_string),
-         call. = FALSE)
+    stop(
+      sprintf(
+        "World Bank API request failed for indicator %s\nmessage: %s\ncategory: %s\nreason: %s \nurl: %s",
+        indicator,
+        error_status$message,
+        error_status$category,
+        error_status$reason,
+        url_string
+      ),
+      call. = FALSE
+    )
   }
 
   if (httr::http_type(get_return) != "application/json") {
@@ -163,5 +188,3 @@ fetch_wb_url_content <- function(url_string, indicator) {
 
   return_json
 }
-
-