+++ R/wbstats-package.R
@@ -1,9 +1,10 @@
-#' wbstats: An R package for searching and downloading data from the World Bank API.
+#' @title Programmatic Access to Data and Statistics from the World Bank API
 #'
-#' The wbstats package provides structured access to data
-#' available from the World Bank API including; support for mutliple languages,
-#' access to annual, quarterly, and monthly data.
+#' @description Search and download data from the World Bank Data API. Includes
+#'  support for mutliple languages, access to annual, quarterly, and monthly
+#'  data.
 #'
-#' @importFrom tibble tibble
+#' @importFrom data.table as.data.table data.table dcast rbindlist set `.SD`
+#'
 #' @name wbstats
 "_PACKAGE"